+++ src/GLFT_Font.h
@@ -83,8 +83,8 @@ class GLFT_Font
 		std::ostream& beginDraw(float x, float y);
 		StreamFlusher endDraw();
 
-        unsigned int calcStringWidth(const char *str, ...) const;
-		unsigned int calcStringWidth(const std::string& str) const;
+		float calcStringWidth(const char *str, ...) const;
+		float calcStringWidth(const std::string& str) const;
 		unsigned int getHeight() const;
 
 	private:
